requirements.txt
alembic==1.7.7
aniso8601==8.0.0
autopep8==1.6.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.2.0
MarkupSafe==1.1.1
pycodestyle==2.8.0
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.2
Werkzeug==1.0.1
zipp==3.8.0
